requirements.txt
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
Faker==37.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
idna==3.10
numpy==2.3.0
packaging==25.0
pandas==2.3.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
matplotlib==3.10
